requirements.txt
telethon==1.36.0
python-dotenv==1.0.1
pydantic==2.7.4
pydantic-settings==2.6.1
pandas==2.2.1
PyYAML==6.0.1
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
uvicorn==0.27.0
fastapi==0.109.0
alembic==1.13.1
ultralytics==8.2.0
opencv-python-headless==4.9.0.80
Pillow==10.2.0
rich==13.7.0
loguru==0.7.2
dbt-postgres==1.7.8
dagster==1.7.16
dagster-webserver==1.7.16
pytest==8.0.0
requests==2.31.0
python-dateutil==2.8.2
textblob==0.17.1
nltk==3.8.1
